// src/components/types/enums/Tag.tsx
export enum Tag {
    Fighter = 'Fighter',
    Mage = 'Mage',
    Bulk_Mage = 'Bulk Mage',
    AP_Assassin = 'Ap Assassin',
    AD_Assassin = 'Ad Assassin',
    Marksman = 'Marksman',
    OnHit = 'On Hit',
    Tank = 'Tank',
    Enchanter_Support = 'Enchanter Support',
    Tank_Support = 'Tank Support',
    Assassin_Support = 'Assassin Support',
    Mage_Support = 'Mage Support',
    OnHit_Mage = 'On Hit Mage',
    Jungle = 'Jungle'
}